+++ changelog.md
@@ -45,6 +45,7 @@ Codion Change Log
 - UiManagerDefaults renamed UIManagerDefaults.
 - FontSize added. DefaultLookAndFeelEnabler now handles updating the font size when enabling.
 - FontSize.updateFontSize() bug fixed, Element Of character got lost and Font instances did not get resized, now uses deriveFont() for FontUIResource instead of instantiation.
+- TextFieldBuilder.SELECT_ALL_ON_FOCUS_GAINED configuration value added.
 ### is.codion.framework.lambda
 - LambdaEntityHandler refactored.